+++ packages/dropdown/package.json
@@ -6,12 +6,11 @@
   "repository": "pluralsight/design-system",
   "author": "jaketrent",
   "keywords": [],
-  "main": "dist/cjs/index.js",
+  "main": "dist/esm/index.js",
   "module": "dist/esm/index.js",
   "scripts": {
-    "build": "run-s build:esm build:cjs build:css",
-    "build:cjs": "tsc --project tsconfig.build.json --module commonjs --target es5 --outDir dist/cjs",
-    "build:css": "build-css --useGlamor -i dist/cjs/css/index.js",
+    "build": "run-s build:esm build:css",
+    "build:css": "build-css --useGlamor -i dist/esm/css/index.js",
     "build:esm": "tsc --project tsconfig.build.json --module es2015 --target es5 --outDir dist/esm",
     "build:watch": "yarn build:esm --watch",
     "storybook": "start-storybook -p 6006",
@@ -19,6 +18,8 @@
     "test:watch": "yarn test --watchAll"
   },
   "gitHead": "7418883d96a76c59a99f86c2b16516ec752bf913",
+  "sideEffects": false,
+  "type": "module",
   "dependencies": {
     "@pluralsight/ps-design-system-core": "^6.9.3",
     "@pluralsight/ps-design-system-halo": "^8.1.4",
