+++ packages/dropdown/src/css/index.ts
@@ -13,7 +13,7 @@ import {
 import { widths as iconWidths } from '@pluralsight/ps-design-system-icon'
 import { names as themeNames } from '@pluralsight/ps-design-system-theme'
 
-import * as vars from '../vars'
+import * as vars from '../vars/index.js'
 
 export default {
   '@keyframes psds-dropdown__menu__keyframes__slide': {
