+++ packages/dropdown/src/index.ts
@@ -1,4 +1,4 @@
-export { default } from './react'
-export * from './vars'
-export { default as css } from './css'
-export { DropdownContext, useDropdown } from './js'
+export { default } from './react/index.js'
+export * from './vars/index.js'
+export { default as css } from './css/index.js'
+export { DropdownContext, useDropdown } from './js/index.js'
