+++ packages/dropdown/src/js/__specs__/index.spec.tsx
@@ -1,8 +1,12 @@
 import { PlaceholderIcon } from '@pluralsight/ps-design-system-icon'
 import React from 'react'
 
-import { getLongestMenuLabel, formatItemId, parseMenuChildren } from '..'
-import Dropdown from '../../react'
+import {
+  getLongestMenuLabel,
+  formatItemId,
+  parseMenuChildren
+} from '../index.js'
+import Dropdown from '../../react/index.js'
 
 describe('#formatItemId', () => {
   it('is menuId-value by default', () => {
