+++ packages/dropdown/src/react/error-icon.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { css } from 'glamor'
 
 import { WarningIcon } from '@pluralsight/ps-design-system-icon'
-import stylesheet from '../css'
+import stylesheet from '../css/index.js'
 
 const styles = css(stylesheet['.psds-dropdown__error'])
