+++ packages/dropdown/src/react/sub-label.tsx
@@ -3,10 +3,10 @@ import {
   names as themeNames,
   useTheme
 } from '@pluralsight/ps-design-system-theme'
-import React from 'react'
 import { compose, css } from 'glamor'
+import React from 'react'
 
-import stylesheet from '../css'
+import stylesheet from '../css/index.js'
 
 const styles = (themeName: ValueOf<typeof themeNames>) => {
   const label = 'psds-dropdown__sub-label'
